feat: [ftest-b00-scenario-boundary-guard-003] - Reject direct internal product invocation from functional tests while preserving approved seams
