+++ contracts/functional-boundary-baseline.json
@@ -10,7 +10,6 @@
     { "path": "tests/functional/guards_batch/same_name_consume_path_ownership_test.go", "sha256": "2a7bdad0d68380f38d613bfadcff855998b12dbe26393a3db7aaf0d0051f5261" },
     { "path": "tests/functional/internal/support/api_server.go", "sha256": "7248cf8ff749f0db662a7d18d8570652ed02f9dbb846d97522bdcd380115df2c" },
     { "path": "tests/functional/internal/support/cmd/browser_api_harness/main.go", "sha256": "2037a06c5d7d7f54b5203d6f23f9702fcd9b988f622e2e1f2715db77362c45e2" },
-    { "path": "tests/functional/providers/construction_contract_test.go", "sha256": "8f40511b883b292d71591016055275b8398c39b273fb132c8cbe3cdb9e8afb7e" },
     { "path": "tests/functional/providers/mock_workers_service_runner_test.go", "sha256": "dd496169f8632c73ec19a19ad105019aa963acaef02dc2ee9351ffd6379a0bad" },
     { "path": "tests/functional/replay_contracts/javascript_failure_contract_test.go", "sha256": "d9810a9685810ad72d08a1de75e189b5c7cbabbd8e1df306a3e53d95b39f79b2" },
     { "path": "tests/functional/replay_contracts/javascript_recording_contract_test.go", "sha256": "dd55dfe16a2fff6dc6c01664012dec3a6edf801b996d432973c58eefdf138588" },
